// client/packages/annotator/src/annotator.types.tsx
export type Maybe<T> = T | null | undefined;
// ...
export interface MetadataInfo {
    key: string;
    value: string;
}

export interface TooltipMetadata {
    [key: number]: MetadataInfo[];
}

export enum Tool {
    SELECT = 'SELECT',
    BBOX = 'BBOX',
    POLYGON = 'POLYGON',
    WAND = 'WAND',
    BRUSH = 'BRUSH',
    ERASER = 'ERASER',
    KEYPOINT = 'KEYPOINT',
    DEXTR = 'DEXTR',
}

// Settings Types
export interface Settings {
    shortcuts: ShortcutsSettings;
    tools: ToolSettings;
}

export interface ShortcutsSettings {
    [key: string]: string;
}

export interface ToolSettings {
    select: Maybe<ToolSettingsSelect>;
    bbox: Maybe<ToolSettingsBBOX>;
    polygon: Maybe<ToolSettingsPolygon>;
    brush: Maybe<ToolSettingsBrush>;
    eraser: Maybe<ToolSettingsBrush>;
    wand: Maybe<ToolSettingsWand>;
}

export interface ToolSettingsSelect {
    tooltipOn: boolean;
}

export interface ToolSettingsBBOX {
    color: string;
}

export interface ToolSettingsPolygon {
    guidanceOn: boolean;
    minDistance: number;
    completeDistance: number;
    colorAuto: boolean;
    colorRadius: number;
    strokeColor: string;
    strokeWidth: number;
}

export interface ToolSettingsBrush {
    color: string;
    radius: number;
}

export interface ToolSettingsWand {
    threshold: number;
    blur: number;
}

// paper canvas types
export interface ImageSize {
    width: number;
    height: number;
}

export interface MouseEvent {
    point: paper.Point;
    item: paper.Item;
    modifiers?: {
        shift?: boolean;
    };
}

export enum DataType {
    INDICATOR = 'INDICATOR',
}

export interface DataIndicator {
    type: DataType.INDICATOR;
}

// paper undo types
export enum ToolEvent {
    POLYGON_ADD_POINT = 'POLYGON_ADD_POINT',
}